Modify, Login working , styles & funcionalities
