--- src/app/auth/loginTest/page.jsx
@@ -1,9 +1,0 @@
-import LoginTest from "@/components/views/auth/Login/LoginTest"
-
-function loginTestPage() {
-  return (
-    <LoginTest/>
-  )
-}
-
-export default loginTestPage
