+++ src/app/auth/loginViejo/page.jsx
@@ -1,0 +1,13 @@
+
+import { Suspense } from "react";
+import LoginViejo from "@/components/views/auth/Login/LoginViejo";
+
+function LoginViejoPage() {
+  return (
+    <Suspense fallback={<div>Loading...</div>}>
+      <Login/>
+    </Suspense>
+  );
+}
+
+export default LoginViejoPage;
