+++ src/components/NavBar/navbar.css
@@ -174,6 +174,7 @@
   font-weight:bold;
   border-radius:10px;
   padding:7px 10px;
+  cursor:pointer;
 }
 
 .nav-main-menu-select-fallback-placeholder{
@@ -226,6 +227,21 @@
 }
 
 .css-1n7v3ny-option:hover {
+  background-color: var(--red) !important;
+  color: white !important;
+}
+
+.css-1n7v3ny-option[aria-selected="true"] {
+  background-color: var(--red) !important;
+  color: white !important;
+}
+
+.nav-main-custom-select__option--is-selected {
+  background-color: var(--red) !important;
+  color: white !important;
+}
+
+.nav-main-custom-select__option:hover {
   background-color: var(--red) !important;
   color: white !important;
 }
