+++ src/components/views/auth/Login/LoginViejo.jsx
@@ -1,0 +1,90 @@
+"use client";
+
+import { useState, useEffect } from "react";
+import { signIn } from "next-auth/react";
+import { useRouter, useSearchParams } from "next/navigation";
+import { useSession } from "next-auth/react";
+import Link from "next/link";
+
+function LoginViejo() {
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const [error, setError] = useState("");
+
+  const router = useRouter();
+  const { status, data: session } = useSession();
+  const searchParams = useSearchParams();
+
+  useEffect(() => {
+    const errorMessage = searchParams.get("error");
+    if (errorMessage) {
+      setError(errorMessage);
+    } else {
+      setError("");
+    }
+  }, [searchParams]);
+
+  useEffect(() => {
+    if (status === "authenticated") {
+      let callbackUrl = searchParams.get("callbackUrl") || "/";
+
+      if (callbackUrl.startsWith("/cart/")) {
+        callbackUrl = `/cart/${session.user.cartId}`;
+      }
+      window.location.assign(callbackUrl);
+    }
+  }, [status, session, router, searchParams]);
+
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    setError("");
+
+    const nextAuthResponse = await signIn("credentials", {
+      email: email,
+      password: password,
+      redirect: false,
+    });
+
+    if (!nextAuthResponse.ok) return setError(nextAuthResponse.error);
+
+    /*  if (nextAuthResponse.ok) {
+      let callbackUrl = searchParams.get("callbackUrl") || "/";
+      console.log(nextAuthResponse);
+      console.log(session);
+      if(callbackUrl.startsWith("/cart/")){
+        callbackUrl = `/cart/${session.user.cartId}`
+      }
+      
+      window.location.assign(callbackUrl);
+    } */
+  };
+
+  return (
+    <div>
+      <form onSubmit={handleSubmit} className="text-black" id="loginViejo-form">
+        <h1>LoginViejo</h1>
+        <input
+          type="email"
+          placeholder="Usuario@gmail.com"
+          name="email"
+          autoComplete="email"
+          required={true}
+          onChange={(e) => setEmail(e.target.value)}
+        />
+        <input
+          type="password"
+          placeholder="********"
+          name="password"
+          autoComplete="current-password"
+          required={true}
+          onChange={(e) => setPassword(e.target.value)}
+        />
+        {error && <p style={{ color: "red" }}>{error}</p>}
+        <button className="bg-white">LoginViejo</button>
+      </form>
+      <p><Link href={"/auth/password-recovery"}>Olvidaste tu contraseña?</Link></p>
+    </div>
+  );
+}
+
+export default LoginViejo;
