+++ src/components/views/auth/Login/login.css
@@ -6,7 +6,7 @@
   align-items:center;
   width:100%;
   height:50vh;
-  margin:150px 0px;
+  margin:90px 0px;
 }
 
 /* Login */
@@ -43,13 +43,24 @@
   justify-content:center;
   align-items:center;
   width:84%;
-  gap:40px;
+  gap:30px;
 }
 
 .login-main-title{
   font-size:25px;
   font-weight:bold;
   color:var(--soft-grey);
+}
+
+.login-main-error{
+  width:100%;
+  margin:5px 0px;
+  padding:3px;
+  color:var(--black);
+  font-size:15px;
+  font-weight:bold;
+  text-align:start;
+  background:rgba(215, 35, 35, 0.4);
 }
 
 
@@ -67,9 +78,10 @@
 .login-main-data input{
   width:100%;
   height:30px;
+  padding:3px;
   background:var(--light-grey);
   border-radius:3px;
-  color:var(--black);
+  color:var(--dark-grey);
   border:1px solid var(--white);
   transition: border 0.2s ease-in-out;
 }
@@ -77,6 +89,19 @@
 .login-main-data input:focus{
   outline:none;
   border:1px solid var(--red);
+}
+
+.login-main-data input::placeholder{
+  font-size:14px;
+  color:var(--soft-grey);
+}
+
+.login-main-forgotpassword{
+  padding-top:10px;
+  text-align: end;
+  font-size:15px;
+  font-weight:bold;
+  color:var(--red);
 }
